requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
efficientnet-pytorch==0.7.1
albumentations==1.3.1
pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.2
websockets==12.0  